requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.6
django-allauth==64.1.0
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jwt==1.3.1
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
setuptools==72.2.0
sqlparse==0.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==7.0.1
